requirements.txt
flask
flasgger
hivecraft==0.3.2
python-dotenv